requirements.txt
selenium
webdriver-manager
pandas
gspread
google-auth
python-dotenv
